+++ NEWS.md
@@ -1,7 +1,7 @@
 # sdcHierarchies 0.19.0
-- [todo] allow non-nested hierarchies
 - major performance improvements by rewriting utility-functions in Rcpp
-- bugfix in `hier_app()` when exporting; 
+- bugfix in `hier_app()` when exporting;
+- feature: `hier_grid()` got new argument `add_default_codes`; if `TRUE`, columns containing standardized codes are automatically added
 
 # sdcHierarchies 0.18.3
 - bugfix when converting hierarchies to sdc-objects.